检索与知识层/faiss_rag/requirements.txt
# ============================================================
# 向量数据库层 - ChromaDB 依赖包
# ============================================================
# 功能：向量数据库构建、存储和检索
# 环境：Python 3.11.13 + CUDA 11.8 + RTX 4060/4090
# 更新时间：2025-10-21
# 状态：✅ 已验证稳定运行
# ============================================================

# ============================================================
# 安装顺序（重要！）
# ============================================================
# 1. pip install torch==2.1.2+cu118 torchvision==0.16.2+cu118 torchaudio==2.1.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# 2. pip install -r requirements.txt
# ============================================================

# ============================================================
# 1. 核心向量数据库（Faiss）
# ============================================================
faiss-cpu==1.7.4  # Faiss向量数据库（CPU版本）
# faiss-gpu==1.7.4  # Faiss向量数据库（GPU版本，如果有NVIDIA GPU）
chromadb==1.2.0
chroma-hnswlib==0.7.3

# ============================================================
# 2. 深度学习和 Embeddings
# ============================================================
# torch==2.1.2+cu118  # 需先单独安装
# torchvision==0.16.2+cu118
# torchaudio==2.1.2+cu118
sentence-transformers==2.3.1
transformers==4.36.2
huggingface-hub==0.24.0
tokenizers==0.15.0
safetensors==0.4.1
accelerate==0.25.0

# ============================================================
# 3. 数据处理
# ============================================================
numpy==1.26.4
pandas==2.0.3
tqdm==4.66.3

# ============================================================
# 4. ChromaDB 必需依赖
# ============================================================
onnxruntime==1.16.3
pydantic==2.12.3
pydantic-core==2.41.4
httpx==0.28.1
httpcore==1.0.9
pybase64==1.4.2
bcrypt==5.0.0
grpcio==1.75.1
pypika==0.48.9
mmh3==5.2.0
overrides==7.7.0
posthog==5.4.0
orjson==3.11.3
pyyaml==6.0.3

# ============================================================
# 5. OpenTelemetry（ChromaDB 监控）
# ============================================================
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0

# ============================================================
# 6. NLP 工具
# ============================================================
nltk==3.8.1
sentencepiece==0.2.1

# ============================================================
# 7. 工具库
# ============================================================
colorama==0.4.6
rich==13.7.1
filelock==3.19.1

# ============================================================
# 使用说明
# ============================================================
# 
# 快速开始：
# 1. conda create -n faiss_rag python=3.11.13
# 2. conda activate faiss_rag
# 3. pip install torch==2.1.2+cu118 torchvision==0.16.2+cu118 torchaudio==2.1.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# 4. pip install -r requirements.txt
# 5. python 构建向量数据库_ChromaDB1.2.py
# 
# 验证：
# python 验证新向量数据库.py
# ============================================================
